+++ tests/benchmarks/test_tic_tac_toe_puzzles.py
@@ -1,4 +1,4 @@
-from decimal import Decimal
+
 import unittest
 
 from src.neuralnet.network import Network
@@ -13,10 +13,10 @@ from src.tictactoe.ai_player import AiPlayer
 from src.tictactoe.board import Board
 
 class Puzzle:
-    def __init__(self, current_state: list[Decimal], desired_move_index: int, desired_move_value: Decimal):
-        self.current_state: list[Decimal] = current_state
+    def __init__(self, current_state: list[float], desired_move_index: int, desired_move_value: float):
+        self.current_state: list[float] = current_state
         self.desired_move_index: int = desired_move_index
-        self.desired_move_value: Decimal = desired_move_value
+        self.desired_move_value: float = desired_move_value
 
 class PuzzleEvaluation(Evaluation):
     def __init__(self, puzzles: list[Puzzle], back_propogate: bool = True):
@@ -35,7 +35,7 @@ class PuzzleEvaluation(Evaluation):
                 network.score += 1
             elif self.back_propogate:
                 network.back_propagate_node_index_and_target(puzzle.desired_move_index, puzzle.desired_move_value)
-                network.back_propagate_node_index_and_target(move, Decimal(0.5))
+                network.back_propagate_node_index_and_target(move, float(0.5))
                 
         network.apply_gradients()
 
@@ -44,88 +44,88 @@ class TestTicTacToePuzzles(unittest.TestCase):
     def setUp(self) -> None:
         self.x_move_puzzles: list[Puzzle] = [
             Puzzle([ # Top Row
-             Decimal(1), Decimal(1), Decimal(0.5), 
-             Decimal(0), Decimal(0.5), Decimal(0.5),
-             Decimal(0.5), Decimal(0), Decimal(0.5)
-            ], 2, Decimal(1)),
+             float(1), float(1), float(0.5), 
+             float(0), float(0.5), float(0.5),
+             float(0.5), float(0), float(0.5)
+            ], 2, float(1)),
             Puzzle([ # Middle Row
-                Decimal(0.5), Decimal(0.5), Decimal(0),
-                Decimal(1), Decimal(0.5), Decimal(1),
-                Decimal(0.5), Decimal(0), Decimal(0.5)
-            ], 4, Decimal(1)),
+                float(0.5), float(0.5), float(0),
+                float(1), float(0.5), float(1),
+                float(0.5), float(0), float(0.5)
+            ], 4, float(1)),
             Puzzle([ # Bottom Row
-                Decimal(0.5), Decimal(0), Decimal(0.5),
-                Decimal(0.5), Decimal(0), Decimal(0.5),
-                Decimal(0.5), Decimal(1), Decimal(1)
-            ], 6, Decimal(1)),
+                float(0.5), float(0), float(0.5),
+                float(0.5), float(0), float(0.5),
+                float(0.5), float(1), float(1)
+            ], 6, float(1)),
             Puzzle([ # Left Column
-                Decimal(1), Decimal(0.5), Decimal(0.5),
-                Decimal(1), Decimal(0), Decimal(0.5),
-                Decimal(0.5), Decimal(0.5), Decimal(0)
-            ], 6, Decimal(1)),
+                float(1), float(0.5), float(0.5),
+                float(1), float(0), float(0.5),
+                float(0.5), float(0.5), float(0)
+            ], 6, float(1)),
             Puzzle([ # Middle Column
-                Decimal(0), Decimal(1), Decimal(0),
-                Decimal(0.5), Decimal(0.5), Decimal(0.5),
-                Decimal(0.5), Decimal(1), Decimal(0.5)
-            ], 4, Decimal(1)), 
+                float(0), float(1), float(0),
+                float(0.5), float(0.5), float(0.5),
+                float(0.5), float(1), float(0.5)
+            ], 4, float(1)), 
             Puzzle([ # Right Column
-                Decimal(0), Decimal(0.5), Decimal(1),
-                Decimal(0), Decimal(0.5), Decimal(1),
-                Decimal(0.5), Decimal(0.5), Decimal(0.5)
-            ], 8, Decimal(1)),
+                float(0), float(0.5), float(1),
+                float(0), float(0.5), float(1),
+                float(0.5), float(0.5), float(0.5)
+            ], 8, float(1)),
             Puzzle([ # Top Left to Bottom Right Diagonal
-                Decimal(0.5), Decimal(0.5), Decimal(0.5),
-                Decimal(0.5), Decimal(1), Decimal(0.5),
-                Decimal(0), Decimal(0), Decimal(1)
-            ], 0, Decimal(1)),
+                float(0.5), float(0.5), float(0.5),
+                float(0.5), float(1), float(0.5),
+                float(0), float(0), float(1)
+            ], 0, float(1)),
             Puzzle([ # Top Right to Bottom Left Diagonal
-                Decimal(0), Decimal(0), Decimal(0.5),
-                Decimal(0.5), Decimal(1), Decimal(0.5),
-                Decimal(1), Decimal(0.5), Decimal(0.5)
-            ], 2, Decimal(0)),
+                float(0), float(0), float(0.5),
+                float(0.5), float(1), float(0.5),
+                float(1), float(0.5), float(0.5)
+            ], 2, float(0)),
         ]
         
         self.o_move_puzzles: list[Puzzle] = [
             Puzzle([ # Top Row
-             Decimal(0), Decimal(0.5), Decimal(0), 
-             Decimal(0.5), Decimal(1), Decimal(0.5),
-             Decimal(1), Decimal(1), Decimal(0.5)
-            ], 1, Decimal(0)),
+             float(0), float(0.5), float(0), 
+             float(0.5), float(1), float(0.5),
+             float(1), float(1), float(0.5)
+            ], 1, float(0)),
             Puzzle([ # Middle Row
-                Decimal(1), Decimal(0.5), Decimal(0.5),
-                Decimal(0), Decimal(0), Decimal(0.5),
-                Decimal(1), Decimal(0.5), Decimal(1)
-            ], 5, Decimal(0)),
+                float(1), float(0.5), float(0.5),
+                float(0), float(0), float(0.5),
+                float(1), float(0.5), float(1)
+            ], 5, float(0)),
             Puzzle([ # Bottom Row
-                Decimal(0.5), Decimal(1), Decimal(0.5),
-                Decimal(1), Decimal(0.5), Decimal(1),
-                Decimal(0), Decimal(0), Decimal(0.5)
-            ], 7, Decimal(0)),
+                float(0.5), float(1), float(0.5),
+                float(1), float(0.5), float(1),
+                float(0), float(0), float(0.5)
+            ], 7, float(0)),
             Puzzle([ # Left Column
-                Decimal(0), Decimal(1), Decimal(0.5),
-                Decimal(0), Decimal(1), Decimal(0.5),
-                Decimal(0.5), Decimal(0.5), Decimal(1)
-            ], 6, Decimal(0)),
+                float(0), float(1), float(0.5),
+                float(0), float(1), float(0.5),
+                float(0.5), float(0.5), float(1)
+            ], 6, float(0)),
             Puzzle([ # Middle Column
-                Decimal(1), Decimal(0), Decimal(0.5),
-                Decimal(0.5), Decimal(0), Decimal(1),
-                Decimal(1), Decimal(0.5), Decimal(0.5)
-            ], 7, Decimal(0)), 
+                float(1), float(0), float(0.5),
+                float(0.5), float(0), float(1),
+                float(1), float(0.5), float(0.5)
+            ], 7, float(0)), 
             Puzzle([ # Right Column
-                Decimal(1), Decimal(1), Decimal(0),
-                Decimal(0.5), Decimal(0.5), Decimal(0),
-                Decimal(0.5), Decimal(1), Decimal(0.5)
-            ], 8, Decimal(0)),
+                float(1), float(1), float(0),
+                float(0.5), float(0.5), float(0),
+                float(0.5), float(1), float(0.5)
+            ], 8, float(0)),
             Puzzle([ # Top Left to Bottom Right Diagonal
-                Decimal(0), Decimal(0.5), Decimal(1),
-                Decimal(1), Decimal(0), Decimal(0.5),
-                Decimal(1), Decimal(0.5), Decimal(0.5)
-            ], 8, Decimal(0)),
+                float(0), float(0.5), float(1),
+                float(1), float(0), float(0.5),
+                float(1), float(0.5), float(0.5)
+            ], 8, float(0)),
             Puzzle([ # Top Right to Bottom Left Diagonal
-                Decimal(1), Decimal(0.5), Decimal(0.5),
-                Decimal(1), Decimal(0), Decimal(0.5),
-                Decimal(0), Decimal(0.5), Decimal(1)
-            ], 2, Decimal(0)),
+                float(1), float(0.5), float(0.5),
+                float(1), float(0), float(0.5),
+                float(0), float(0.5), float(1)
+            ], 2, float(0)),
         ]
     
     """
